refactor(control): move parseFlags + ensureAdminUser to sibling files (#157 slice 5)
Fifth slice. Pure file-shuffling — zero behaviour change.

- parseFlags + applyEnvOverrides + clampDurations + mustValidateConfig
  \xe2\x86\x92 cmd/control/flags.go (129 LOC)
- ensureAdminUser + maskDatabaseURL \xe2\x86\x92 cmd/control/admin_user.go (94 LOC)

main.go: 528 \xe2\x86\x92 348 LOC (-180). Cumulative from baseline 1022: -674 LOC (-66%).

The flag-parsing function was also split into three named phases
(applyEnvOverrides, clampDurations, mustValidateConfig) so each is
individually testable and the parseFlags shape reads as a sequence
of phases rather than 100 LOC of inline imperative wiring.

Stacked on top of #228 (slice 3) + the local slice 4 branch (Valkey
subsystem) \xe2\x80\x94 will rebase as those land.

Refs manchtools/power-manage-server#157 (audit F043). main() is now
~150 LOC of subsystem orchestration; one more pass on the public/
internal HTTP listener startup goroutines would land it at the
issue's <300 LOC bar.
